refactor: move CreateTransactionJSON to transaction.go :hammer:
