+++ components/ledger/internal/adapters/http/in/transaction.go
@@ -5,11 +5,60 @@
 package in
 
 import (
+	libCommons "github.com/LerianStudio/lib-commons/v4/commons"
+
+	"github.com/LerianStudio/midaz/v3/components/ledger/internal/adapters/postgres/transaction"
 	"github.com/LerianStudio/midaz/v3/components/ledger/internal/services/command"
 	"github.com/LerianStudio/midaz/v3/components/ledger/internal/services/query"
+	"github.com/LerianStudio/midaz/v3/pkg/constant"
+
+	"github.com/gofiber/fiber/v2"
 )
 
 type TransactionHandler struct {
 	Command *command.UseCase
 	Query   *query.UseCase
+}
+
+// CreateTransactionJSON method that create transaction using JSON
+//
+//	@Summary		Create a Transaction using JSON
+//	@Description	Create a Transaction with the input payload
+//	@Tags			Transactions
+//	@Accept			json
+//	@Produce		json
+//	@Param			Authorization	header		string						true	"Authorization Bearer Token"
+//	@Param			X-Request-Id	header		string						false	"Request ID"
+//	@Param			organization_id	path		string						true	"Organization ID"
+//	@Param			ledger_id		path		string						true	"Ledger ID"
+//	@Param			transaction		body		transaction.CreateTransactionSwaggerModel	true	"Transaction Input"
+//	@Success		201				{object}	Transaction
+//	@Failure		400				{object}	mmodel.Error	"Invalid input, validation errors"
+//	@Failure		401				{object}	mmodel.Error	"Unauthorized access"
+//	@Failure		403				{object}	mmodel.Error	"Forbidden access"
+//	@Failure		422				{object}	mmodel.Error	"Unprocessable Entity, validation errors"
+//	@Failure		500				{object}	mmodel.Error	"Internal server error"
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/transactions/json [post]
+func (handler *TransactionHandler) CreateTransactionJSON(p any, c *fiber.Ctx) error {
+	ctx := c.UserContext()
+
+	logger, tracer, _, _ := libCommons.NewTrackingFromContext(ctx)
+
+	ctx, span := tracer.Start(ctx, "handler.create_transaction")
+	defer span.End()
+
+	c.SetUserContext(ctx)
+
+	input := p.(*transaction.CreateTransactionInput)
+	transactionInput := input.BuildTransaction()
+	logSafePayload(ctx, logger, "Request to create a transaction", transactionInput)
+
+	transactionStatus := constant.CREATED
+	if transactionInput.Pending {
+		transactionStatus = constant.PENDING
+	}
+
+	recordSafePayloadAttributes(span, transactionInput)
+
+	return handler.createTransaction(c, *transactionInput, transactionStatus)
 }
